optim

+++ src/attacker/rollout_helper/common/base_rollout_helper.py
@@ -23,7 +23,7 @@ class BaseRolloutHelper(RolloutHelper):
         self.baseline_obs_dist = baseline_obs_len
 
         actions = [action for action in range(n_actions)]
-        self.onehot_action = F.one_hot(torch.tensor(actions), num_classes=self.n_actions)
+        self.onehot_action = F.one_hot(torch.tensor(actions), num_classes=self.n_actions).float()
         self.action_enumeration = list(product(actions, repeat=self.action_enum_len))
 
         self.frame_cycler = FrameCycler()
